Predicate for counting a persistence corrected; rg_corrections function placed outside calc_dist_matrix function; rg_corrections takes in a list of selections and returns a numpy array; new user-inputted argument for choosing correction function; null correction function added and is default

+++ libinteract/clibinteract.h
@@ -25,6 +25,6 @@ double ed(double*, double*, int, int);
 int potential_distances(double*, int, int, int, double*);
 int triangular_distmatrix(double*, int, int, double, long*, double*);
 int square_distmatrix(double*, double*, int, int, int, double, long*);
-int triangular_mindist(double*, int, int, long*, double, double*);
+int triangular_mindist(double*, int, int, long*, double, long*);
 int square_mindist(double*, double*, int, int, int, long*, long*, double, long*);
